+++ MainProject/Injecter/ScopeStore.cs
@@ -1,27 +1,29 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using System;
-using System.Collections.Generic;
+using System.Collections.Concurrent;
 
 namespace Injecter
 {
-    public sealed class ScopeStore : IScopeStore
+    public sealed class ScopeStore : IScopeStore, IDisposable
     {
         private readonly IServiceProvider _serviceProvider;
-        private readonly Dictionary<object, IServiceScope> _scopes = new();
+        private readonly ConcurrentDictionary<object, IServiceScope> _scopes = new();
+
+        public int NumberOfScopes => _scopes.Count;
 
         public ScopeStore(IServiceProvider serviceProvider) => _serviceProvider = serviceProvider;
 
-        public IServiceScope CreateScope<T>(T owner) where T : notnull
+        public IServiceScope CreateScope(object owner)
         {
             if (owner == null) throw new ArgumentNullException(nameof(owner));
 
             var scope = _serviceProvider.CreateScope();
-            _scopes.Add(owner, scope);
+            _scopes.TryAdd(owner, scope);
 
             return scope;
         }
 
-        public IServiceScope? GetScope<T>(T owner) where T : notnull
+        public IServiceScope? GetScope(object owner)
         {
             if (owner == null) throw new ArgumentNullException(nameof(owner));
 
@@ -30,7 +32,7 @@ namespace Injecter
             return scope;
         }
 
-        public void DisposeScope<T>(T owner) where T : notnull
+        public void DisposeScope(object owner)
         {
             if (owner == null) throw new ArgumentNullException(nameof(owner));
 
@@ -38,7 +40,7 @@ namespace Injecter
 
             scope.Dispose();
 
-            _scopes.Remove(owner);
+            _scopes.TryRemove(owner, out _);
         }
 
         public void ClearAllScopes()
@@ -50,5 +52,7 @@ namespace Injecter
 
             _scopes.Clear();
         }
+
+        public void Dispose() => ClearAllScopes();
     }
 }