+++ UnityProject/Injecter.Unity/Assets/Injecter.Unity.Tests/Arrange/ComplexSceneInjectedProperly/ComplexConcrete.cs
@@ -4,10 +4,10 @@ namespace Injecter.Unity.Tests.Arrange.ComplexSceneInjectedProperly
 {
     public sealed class ComplexConcrete : ComplexBase
     {
-        private ITestService1 _testService1;
+        private TestService _testService1;
 
         [Inject]
-        public void Initialize(ITestService1 testService1) => _testService1 = testService1;
+        public void Initialize(TestService testService1) => _testService1 = testService1;
 
         public void AssertConcreteInjected() => Assert.NotNull(_testService1);
     }