+++ UnityProject/Injecter.Unity/Assets/Injecter.Unity.Tests/Arrange/ComplexSceneInjectedProperly/TestService1.cs
@@ -1,6 +1,0 @@
-﻿namespace Injecter.Unity.Tests.Arrange.ComplexSceneInjectedProperly
-{
-    public sealed class TestService1 : ITestService1
-    {
-    }
-}